Move i18n logic to separate directory
