+++ src/i18n/__init__.py
@@ -1,0 +1,1 @@
+from .i18n import SUPPORTED_LANGUAGES, initialize_locales, translate
